Updated fortun-pony-motd.sh for Mac
Checks to see if this is Darwin, or if update-motd.d does not exist. If
so, it will output the fortune-pony stuff to the user's bash_profile

+++ fortune-pony-motd.sh
@@ -1,36 +1,50 @@
 #!/bin/bash
 
-if [[ `/usr/bin/id -u` -ne 0 ]]; then
-  echo "Try sudo ./fortunepony-motd.sh"
-else
-  echo "Configuring ponysay motd :D"
-  if [[ ! `command -v ponysay` ]]; then
-    echo "Ponysay is not installed. Try this:"
-    echo
-    echo "cd $HOME"
-    echo "git clone https://github.com/erkin/ponysay.git"
-    echo "cd ponysay"
-    echo "sudo make"
-    echo
-  else
-    read -r -d '' fortunepony <<'EOF'
-#!/bin/bash
+read -r -d '' fortunepony <<'EOF'
 echo
 fortune | sed -e 's/america/amareica/g' -e 's/America/Amareica/g' -e 's/anybody/anypony/g' -e 's/Anybody/Anypony/g' -e 's/anyone/anypony/g' -e 's/Anyone/Anypony/g' -e 's/boy/colt/g' -e 's/Boy/Colt/g' -e 's/butthurt/saddle-sore/g' -e 's/Butthurt/Saddlesore/g' -e 's/children/foals/g' -e 's/Children/Foals/g' -e 's/child/foal/g' -e 's/Child/Foal/g' -e 's/cowboy/cowpony/g' -e 's/Cowboy/Cowpony/g' -e 's/cowgirl/cowpony/g' -e 's/Cowgirl/Cowpony/g' -e 's/gentlemen/gentlecolts/g' -e 's/Gentlemen/Gentlecolts/g' -e 's/everybody/everypony/g' -e 's/Everybody/Everypony/g' -e 's/everyone/everypony/g' -e 's/Everyone/Everypony/g' -e 's/feet/hooves/g' -e 's/Feet/Hooves/g' -e 's/folks/foalks/g' -e 's/Folks/Foalks/g' -e 's/fool/foal/g' -e 's/Fool/Foal/g' -e 's/foot/hoof/g' -e 's/Foot/Hoof/g' -e 's/girls/fillies/g' -e 's/Girls/Fillies/g' -e 's/girl/filly/g' -e 's/Girl/Filly/g' -e 's/halloween/nightmare night/g' -e 's/Halloween/Nightmare Night/g' -e 's/hands/hooves/g' -e 's/Hands/Hooves/g' -e 's/handed/hooved/g' -e 's/Handed/Hooved/g' -e 's/hand/hoof/g' -e 's/Hand/Hoof/g' -e 's/\bhey\b/hay/g' -e 's/\bHey\b/Hay/g' -e 's/humans/ponies/g' -e 's/Humans/Ponies/g' -e 's/human/pony/g' -e 's/Human/Pony/g' -e 's/ladies/fillies/g' -e 's/Ladies/Fillies/g' -e 's/main/mane/g' -e 's/Main/Mane/g' -e 's/woman/mare/g' -e 's/Woman/Mare/g' -e 's/women/mares/g' -e 's/Women/Mares/g' -e 's/\bman\b/stallion\ /g' -e 's/\bMan\b/Stallion/g' -e 's/\bmen\b/stallions/g' -e 's/\bMen\b/Stallions/g' -e 's/no\ one\ else/no\ pony\ else/g' -e 's/No\ one\ else/No\ pony\ else/g' -e 's/nobody/nopony/g' -e 's/Nobody/Nopony/g' -e 's/people/ponies/g' -e 's/People/Ponies/g' -e 's/person/pony/g' -e 's/Person/Pony/g' -e 's/philadelphia/fillydelphia/g' -e 's/Philadelphia/Fillydelphia/g' -e 's/somebody/somepony/g' -e 's/Somebody/Somepony/g' -e 's/tattoo/cutie\ mark/g' -e 's/Tattoo/Cutie\ mark/g' -e 's/the\ heck/the\ hay/g' -e 's/The\ heck/The hay/g' -e 's/the\ hell/the\ hay/g' -e 's/The\ Hell/The\ Hay/g' -e 's/troll/parasprite/g' -e 's/Troll/Parasprite/g' | ponysay
 EOF
 
-    if command -v cowsay >/dev/null 2>&1 && command -v fortune >/dev/null 2>&1; then
-      [[ `command -v cowsay` == "/usr/bin/cowsay" ]] || ln -s `which cowsay` /usr/bin/cowsay
-      [[ `command -v fortune` == "/usr/bin/fortune" ]] || ln -s `which fortune` /usr/bin/fortune
-
-      echo "$fortunepony" | tee /etc/update-motd.d/51-fortune-pony >/dev/null 2>&1 &
-      if [[ -f /etc/update-motd.d/51-fortune-pony ]]; then
-        chmod 755 /etc/update-motd.d/51-fortune-pony
-      fi
+# try to update the motd, or else add it to the bottom of the .bash_profile
+if [[ ! `uname -s` = 'Darwin' && -f /etc/update-motd.d ]]; then
+  # most likely on linux
+  if [[ `/usr/bin/id -u` -ne 0 ]]; then
+    echo "Try sudo ./fortunepony-motd.sh"
+  else
+    echo "Configuring ponysay motd :D"
+    if [[ ! `command -v ponysay` ]]; then
+      echo "Ponysay is not installed. Try this:"
+      echo
+      echo "cd $HOME"
+      echo "git clone https://github.com/erkin/ponysay.git"
+      echo "cd ponysay"
+      echo "sudo make"
+      echo
     else
-      `command -v cowsay >/dev/null 2>&1` || echo "Cowsay is not installed. Try this: sudo apt-get install cowsay"
-      `command -v fortune >/dev/null 2>&1` || echo "Fortune is not installed. Try this: sudo apt-get install fortune"
-      echo "There was a problem configuring ponysay motd."
+      if command -v cowsay >/dev/null 2>&1 && command -v fortune >/dev/null 2>&1; then
+        # need these because we're running as root and root doesn't have /usr/games in his path
+        [[ `command -v cowsay` == "/usr/bin/cowsay" ]] || ln -s `which cowsay` /usr/bin/cowsay
+        [[ `command -v fortune` == "/usr/bin/fortune" ]] || ln -s `which fortune` /usr/bin/fortune
+  
+        echo "#!/bin/bash" | tee /etc/update-motd.d/51-fortune-pony >/dev/null 2>&1 &
+        echo "$fortunepony" | tee /etc/update-motd.d/51-fortune-pony >/dev/null 2>&1 &
+        if [[ -f /etc/update-motd.d/51-fortune-pony ]]; then
+          chmod 755 /etc/update-motd.d/51-fortune-pony
+        fi
+      else
+        `command -v cowsay >/dev/null 2>&1` || echo "Cowsay is not installed. Try this: sudo apt-get install cowsay"
+        `command -v fortune >/dev/null 2>&1` || echo "Fortune is not installed. Try this: sudo apt-get install fortune"
+        echo "There was a problem configuring ponysay motd."
+      fi
     fi
+  fi
+else
+  # this is a Mac (or something without update-motd.d, but we'll assume Mac)
+  if command -v cowsay >/dev/null 2>&1 && command -v fortune >/dev/null 2>&1; then
+    echo "$fortunepony" | tee -a $HOME/.bash_profile >/dev/null 2>&1
+  else
+    `command -v cowsay >/dev/null 2>&1` || echo "Cowsay is not installed. Try this: sudo brew install cowsay"
+    `command -v fortune >/dev/null 2>&1` || echo "Fortune is not installed. Try this: sudo brew install fortune"
+    echo "There was a problem configuring ponysay motd."
   fi
 fi
